% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME2_UserLevelFunctions.R
\name{AddCarbonTaxToPrices}
\alias{AddCarbonTaxToPrices}
\title{Title AddCarbonTaxToPrices}
\usage{
AddCarbonTaxToPrices(D)
}
\arguments{
\item{D}{A tibble with matrix-columns object}
}
\value{
D A tibble with matrix-columns object
}
\description{
Title AddCarbonTaxToPrices
}
