% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ApplyPowerModelToSpecificYear}
\alias{ApplyPowerModelToSpecificYear}
\title{ApplyPowerModelToSpecificYear}
\usage{
ApplyPowerModelToSpecificYear(
  CurrentYearTemp,
  CTMaxRate = 0,
  MainPowerDataTable = MainPowerLargeDataTable_,
  BaseYear__ = BaseYear_,
  NumberOfYears__ = NumberOfYears_,
  RetirementProportion__ = RetirementProportion_
)
}
\description{
ApplyPowerModelToSpecificYear
}
