% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T01_ConfigSimplifiedCPAT.R
\name{BaselineInputs}
\alias{BaselineInputs}
\title{Baseline Inputs:
This function reads some pre-loaded scenarios to test CPAT}
\usage{
BaselineInputs(BaseL)
}
\arguments{
\item{BaseL}{BaseList included in cpatr with templates}
}
\description{
Baseline Inputs:
This function reads some pre-loaded scenarios to test CPAT
}
