% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BaseList}
\alias{BaseList}
\title{BaseList
A list providing information on base variables not currently modifyble by the user.
These variables determine the dimensions of the data structures, as they provide starting year, end year,
number of countries, among others.
The list also includes an empty template matrix that is used to create the new matrix columns whenever needed.}
\format{
List
}
\usage{
BaseList
}
\description{
BaseList
A list providing information on base variables not currently modifyble by the user.
These variables determine the dimensions of the data structures, as they provide starting year, end year,
number of countries, among others.
The list also includes an empty template matrix that is used to create the new matrix columns whenever needed.
}
\keyword{datasets}
