% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{BuildMainSegmentedDataTable}
\alias{BuildMainSegmentedDataTable}
\title{BuildMainSegmentedDataTable}
\usage{
BuildMainSegmentedDataTable(
  Base_,
  Elasticities_,
  Fuels_,
  CPATEnergyBalancesBaseYear_,
  GDPRelativeToBase_,
  FuelPrices_,
  CarbonTaxTrajectoryForm_
)
}
\arguments{
\item{CountryList_}{List of Countries}
}
\value{
tibble
}
\description{
Create the MainSegmentedDataTable
}
