% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T04_DataInclExogProj.R
\name{CompleteYears}
\alias{CompleteYears}
\title{Include years missing in a dataframe}
\usage{
CompleteYears(
  TheDF,
  BaseL = BaseList,
  HistY = "with_NA",
  ProjY = "ConstantGrowth"
)
}
\arguments{
\item{TheDF}{Dataframe to modify}

\item{BaseL}{List with global parameters and templates to reconstruct matrices consistent with DD}

\item{HistY}{Rule used to complete the data for historical missing years. (Current option: 'with_NA')}

\item{ProjY}{Rule used to complete the data for projected years (Current option: 'ConstantGrowth')}
}
\description{
Include years missing in a dataframe
}
