% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T03_SupportFunctions.R
\name{Convert_D_to_T}
\alias{Convert_D_to_T}
\title{Convert matrix column into longer tibble}
\usage{
Convert_D_to_T(DD, ColCateg = "Year", ReplColCateg = "Constant")
}
\arguments{
\item{DD}{Tibble using matrix columns (D matrix) to be transformed into longer format}

\item{ColCateg}{Name of the variable that will store the current column names of DD}

\item{ReplColCateg}{Entry to replace the ColCateg for vectors of DD that are not matrix columns}
}
\description{
Convert matrix column into longer tibble
}
