% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T03_SupportFunctions.R
\name{Convert_T_to_D}
\alias{Convert_T_to_D}
\title{Convert dataframe in long format into tibble with matrix columns}
\usage{
Convert_T_to_D(
  TT,
  ColVarNames = "Variable",
  NamesToWide = "Year",
  ValuesToWide = "Value"
)
}
\arguments{
\item{TT}{Tibble in longer format, to convert into a tibble with matrix columns}

\item{ColVarNames}{Column in TT where the variable names are stored}

\item{NamesToWide}{Column in TT from which to read the names for the columns within each matrix column}

\item{ValuesToWide}{Column in TT where the variable values are stored}
}
\description{
Convert dataframe in long format into tibble with matrix columns
}
