% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TibbleMatrixFunctions.R
\name{convert_tibble_matrix_to_tibble}
\alias{convert_tibble_matrix_to_tibble}
\title{Title}
\usage{
convert_tibble_matrix_to_tibble(data)
}
\arguments{
\item{data}{a tibble-matrix object}
}
\value{

}
\description{
Title
}
