% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ConvertTibbleWithIDCols}
\alias{ConvertTibbleWithIDCols}
\title{ConvertTibbleWithIDCols}
\usage{
ConvertTibbleWithIDCols(tb, idcols = 1L, conversion = 1)
}
\arguments{
\item{tb}{Parameter}

\item{idcols}{Parameter}

\item{conversion}{Parameter}
}
\description{
ConvertTibbleWithIDCols
}
