% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T03_SupportFunctions.R
\name{CreateCP}
\alias{CreateCP}
\title{Carbon price trajectory}
\usage{
CreateCP(
  StartCPYear = 2023,
  TargetCPYear = 2030,
  StartCP = 60,
  TargetCP = 100,
  TemplMat
)
}
\arguments{
\item{StartCPYear}{Starting year for the carbon price}

\item{TargetCPYear}{Year for the carbon price to reach the desired target level}

\item{StartCP}{Starting value for the carbon price}

\item{TargetCP}{Target value for the carbon price to reach by TargetCPYear}

\item{TemplMat}{Template matrix of 0s, with number of rows and columns consistent with the D matrix.}
}
\description{
Carbon price trajectory
}
