% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DB_AllCountries}
\alias{DB_AllCountries}
\title{DB_AllCountries
An array including matrix columns for all countries and years selected by default.
This dataset contains the historical information for key variables that do not depend on user-defined parameters
and provides space ready-to-be-filled for forecasts.
@format Tibble, array. It includes single columns to be read as column vectors, and also others as matrix columns.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52920 rows and 32 columns.
}
\usage{
DB_AllCountries
}
\description{
DB_AllCountries
An array including matrix columns for all countries and years selected by default.
This dataset contains the historical information for key variables that do not depend on user-defined parameters
and provides space ready-to-be-filled for forecasts.
@format Tibble, array. It includes single columns to be read as column vectors, and also others as matrix columns.
}
\keyword{datasets}
