% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME2_UserLevelFunctions.R
\name{EnergyUseTminusonetoT}
\alias{EnergyUseTminusonetoT}
\title{Title}
\usage{
EnergyUseTminusonetoT(D = D, InitY = StartYear, EndY = EndYear, i = 1)
}
\arguments{
\item{i}{}
}
\value{

}
\description{
Title
}
