% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T05_DatabaseFunctions.R
\name{fetch_table}
\alias{fetch_table}
\title{#' Load Data from SQL
#'
#' @param con SQL Connection Object
#' @param TableID Table ID as saved to SQL DB
#'
#' @return
#' @export
#' @import tidyr dplyr purrr RPostgres DBI
#'
#' @examples}
\usage{
fetch_table(con, TableID)
}
\description{
#' Load Data from SQL
#'
#' @param con SQL Connection Object
#' @param TableID Table ID as saved to SQL DB
#'
#' @return
#' @export
#' @import tidyr dplyr purrr RPostgres DBI
#'
#' @examples
}
