% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T03_SupportFunctions.R
\name{FocusCountry}
\alias{FocusCountry}
\title{This function:
\itemize{
\item Filters historic data to keep only that relevant for the selected countries
\item It adapts templates and elements from the BaseList to match the new dimensions needed
}}
\usage{
FocusCountry(DD, FullBaseL, SelectedCountryList, LocalUserScen)
}
\arguments{
\item{DD}{Tibble using matrix columns (D matrix) to be filtered}

\item{FullBaseL}{List with pre-defined defaults and templates}

\item{SelectedCountryList}{List of selected countries}

\item{LocalUserScen}{List of all inputs from user, except the Countries selected}
}
\description{
This function:
\itemize{
\item Filters historic data to keep only that relevant for the selected countries
\item It adapts templates and elements from the BaseList to match the new dimensions needed
}
}
