% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T02_CPATCoreFunctions.R
\name{ForecastDomPrices}
\alias{ForecastDomPrices}
\title{Domestic prices forecasting}
\usage{
ForecastDomPrices(
  DD,
  BaseL = BaseList,
  PolInputs = MTI[[ss]],
  ScenName = ss,
  Year
)
}
\arguments{
\item{DD}{Tibble with historical information and required structure in matrix column format (D-matrix)}

\item{BaseL}{List with global parameters and templates to reconstruct matrices consistent with DD}

\item{PolInputs}{List of policy details to consider for the current scenario}

\item{ScenName}{Scenario name, consistent with the names used in PolInputs}

\item{Year}{Year of simulation. The process builds the forecast by yearly time-steps}
}
\description{
Domestic prices forecasting
}
