% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T05_DatabaseFunctions.R
\name{insert_table}
\alias{insert_table}
\title{#' Insert Table to SQL
#'
#' @param con SQL Connection Object
#' @param data Data in tibble or data.frame format
#' @param TableID Only Lowercase Letters, Numbers, and Underscores
#' @param TableName Full Name of Table
#' @param TableCategory Category or Project Name
#' @param Date Date R object, as.Date('yyyy-mm-dd')
#' @param Comments ...
#'
#' @return
#' @import tidyr dplyr purrr RPostgres DBI
#'
#' @examples
#' @export}
\usage{
insert_table(con, data, TableID, TableName, TableCategory, Date, Comments)
}
\description{
#' Insert Table to SQL
#'
#' @param con SQL Connection Object
#' @param data Data in tibble or data.frame format
#' @param TableID Only Lowercase Letters, Numbers, and Underscores
#' @param TableName Full Name of Table
#' @param TableCategory Category or Project Name
#' @param Date Date R object, as.Date('yyyy-mm-dd')
#' @param Comments ...
#'
#' @return
#' @import tidyr dplyr purrr RPostgres DBI
#'
#' @examples
#' @export
}
