% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TibbleMatrixFunctions.R
\name{is.tibble.matrix}
\alias{is.tibble.matrix}
\title{Verify if tibble contains matrix-columns}
\usage{
is.tibble.matrix(data)
}
\arguments{
\item{data}{A tibble object}
}
\value{
An object of class \code{logical}, TRUE or FALSE
}
\description{
With the \code{is.tibble.matrix()} function we check every column of the tibble object
and detect if any of them contain a matrix-column

More documentation on matrix-columns can be found at \url{https://www.ericrscott.com/post/matrix-columns/}
}
\examples{
is.tibble.matrix(mtcars)
data("Dmatrix")
is.tibble.matrix(Dmatrix)
}
