% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T02_CPATCoreFunctions.R
\name{MitEQ}
\alias{MitEQ}
\title{Energy Use Equations}
\usage{
MitEQ(DD, BaseL, PolInputs = MTI[[ss]], Year)
}
\arguments{
\item{DD}{Tibble with historical information and required structure in matrix column format (D-matrix)}

\item{BaseL}{List with global parameters and templates to reconstruct matrices consistent with DD}

\item{PolInputs}{List of policy details to consider for the current scenario}

\item{Year}{Year of simulation. The process builds the forecast by yearly time-steps}
}
\description{
Energy Use Equations
}
