% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{OverallCoremodel}
\alias{OverallCoremodel}
\title{OverallCoremodel}
\usage{
OverallCoremodel(
  CountryList = gCountryList,
  DoPower = gDoPower,
  BaseYear = gBaseYear,
  NumberOfYears = gNumberOfYears,
  AnalysisEndYear = gAnalysisEndYear,
  EndYear = gEndYear,
  RetirementProportion = gRetirementProportion,
  CarbonTaxTrajectoryForm = gCarbonTaxTrajectoryForm,
  CTRange = gCTRange
)
}
\description{
Run OverallCoremodel
}
