% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T04_DataInclExogProj.R
\name{PrepareGDPRelativeToBase}
\alias{PrepareGDPRelativeToBase}
\title{GDP relative to base year}
\usage{
PrepareGDPRelativeToBase(DD, BaseL, RawGDPRelativeToBase)
}
\arguments{
\item{DD}{Tibble with variables in matrix column format (D matrix)}

\item{BaseL}{List with global parameters and templates to reconstruct matrices consistent with DD}

\item{RawGDPRelativeToBase}{}
}
\description{
GDP relative to base year
}
