% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T04_DataInclExogProj.R
\name{PrepocessIntPricesList}
\alias{PrepocessIntPricesList}
\title{Historical and projected international energy prices}
\usage{
PrepocessIntPricesList(
  BaseL,
  RInputs_InternationalPrices_RegionAssumptions,
  RInputs_InternationalPrices_RegionMarket,
  RInputs_InternationalPrices_IntPrices
)
}
\arguments{
\item{BaseL}{List with global parameters and templates to reconstruct matrices consistent with DD}
}
\description{
Historical and projected international energy prices
}
