% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RawGDPRelativeToBase}
\alias{RawGDPRelativeToBase}
\title{RawGDPRelativeToBase
Dataset with information on the real GDP in each year over that of the base year. Used to compute Energy Consumption}
\format{
tibble
}
\usage{
RawGDPRelativeToBase
}
\description{
RawGDPRelativeToBase
Dataset with information on the real GDP in each year over that of the base year. Used to compute Energy Consumption
}
\keyword{datasets}
