% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RInputs_InternationalPrices_IntPrices}
\alias{RInputs_InternationalPrices_IntPrices}
\title{RInputs_InternationalPrices_IntPrices}
\format{
tibble
}
\usage{
RInputs_InternationalPrices_IntPrices
}
\description{
RInputs_InternationalPrices_IntPrices
}
\keyword{datasets}
