% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RInputs_RawGDPRelativeToBase}
\alias{RInputs_RawGDPRelativeToBase}
\title{RInputs_RawGDPRelativeToBase
Dataset with information on the real GDP in each year over that of the base year. Used to compute Energy Consumption}
\format{
tibble
}
\usage{
RInputs_RawGDPRelativeToBase
}
\description{
RInputs_RawGDPRelativeToBase
Dataset with information on the real GDP in each year over that of the base year. Used to compute Energy Consumption
}
\keyword{datasets}
