% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T00_MainSimplifiedCPAT.R
\name{SimpleCPAT}
\alias{SimpleCPAT}
\title{CPAT R v0.0.0}
\usage{
SimpleCPAT(FullHistoricDataset, FullBaseList, UserScen, CountryList)
}
\arguments{
\item{FullHistoricDataset}{Default dataset with historical information and required format}

\item{FullBaseList}{List with global parameters and templates to reconstruct matrices consistent with DD}

\item{UserScen}{Scenario parameters provided by the user}

\item{CountryList}{List of countries selected by the user}
}
\description{
CPAT R v0.0.0
}
