% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TibbleMatrixFunctions.R
\name{tibble_to_tibble_matrix}
\alias{tibble_to_tibble_matrix}
\title{Title}
\usage{
tibble_to_tibble_matrix(data)
}
\arguments{
\item{data}{a tibble to convert to tibble-matrix}
}
\value{

}
\description{
Title
}
