% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{UpdateCoreData}
\alias{UpdateCoreData}
\title{UpdateCoreData}
\usage{
UpdateCoreData(MainSegmentedDataTableStyleInput, CTMaxRate = CTMaxRate)
}
\arguments{
\item{MainSegmentedDataTableStyleInput}{Parameter}

\item{CTMaxRate}{Parameter}
}
\description{
Update Core Data
}
