% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME2_UserLevelFunctions.R
\name{UpdatesAllDataPoints}
\alias{UpdatesAllDataPoints}
\title{Title}
\usage{
UpdatesAllDataPoints(Dint = D, InitY = StartYear, EndY = EndYear)

UpdatesAllDataPoints(Dint = D, InitY = StartYear, EndY = EndYear)
}
\arguments{
\item{EndY}{}
}
\value{



}
\description{
Title

Title
}
