% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{BuildBaseTable}
\alias{BuildBaseTable}
\title{Build Base Table}
\usage{
BuildBaseTable(
  CountryList_,
  SubsectorList_,
  FuelList_,
  YearList_,
  CTRange_,
  BaseYear_,
  SectorSubsectorLookup_,
  PowerFuelTypes_
)
}
\arguments{
\item{CountryList_}{List of Countries}

\item{SubsectorList_}{Parameter}

\item{FuelList_}{Parameter}

\item{YearList_}{Parameter}

\item{CTRange_}{Parameter}

\item{BaseYear_}{Parameter}

\item{SectorSubsectorLookup_}{Parameter}

\item{PowerFuelTypes_}{Parameter}
}
\value{
tibble
}
\description{
Create the base table.
}
