% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{BuildMainSegmentedDataTable}
\alias{BuildMainSegmentedDataTable}
\title{BuildMainSegmentedDataTable}
\usage{
BuildMainSegmentedDataTable(
  Base_,
  Elasticities_,
  Fuels_,
  CPATEnergyBalancesBaseYear_,
  GDPRelativeToBase_,
  FuelPrices_,
  CarbonTaxTrajectoryForm_
)
}
\arguments{
\item{Base_}{Parameter}

\item{Elasticities_}{Parameter}

\item{Fuels_}{Parameter}

\item{CPATEnergyBalancesBaseYear_}{Parameter}

\item{GDPRelativeToBase_}{Parameter}

\item{FuelPrices_}{Parameter}

\item{CarbonTaxTrajectoryForm_}{Parameter}
}
\value{
tibble
}
\description{
Create the MainSegmentedDataTable
}
