% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{OverallCoremodel}
\alias{OverallCoremodel}
\title{OverallCoremodel}
\usage{
OverallCoremodel(
  CountryList = gCountryList,
  DoPower = gDoPower,
  BaseYear = gBaseYear,
  NumberOfYears = gNumberOfYears,
  AnalysisEndYear = gAnalysisEndYear,
  EndYear = gEndYear,
  RetirementProportion = gRetirementProportion,
  CarbonTaxTrajectoryForm = gCarbonTaxTrajectoryForm,
  CTRange = gCTRange
)
}
\arguments{
\item{CountryList}{gCountryList}

\item{DoPower}{gDoPower}

\item{BaseYear}{gBaseYear}

\item{NumberOfYears}{gNumberOfYears}

\item{AnalysisEndYear}{gAnalysisEndYear}

\item{EndYear}{gEndYear}

\item{RetirementProportion}{gRetirementProportion}

\item{CarbonTaxTrajectoryForm}{gCarbonTaxTrajectoryForm}

\item{CTRange}{gCTRange}
}
\description{
Run OverallCoremodel
}
