% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RInputs_InternationalPrices_RegionMarket}
\alias{RInputs_InternationalPrices_RegionMarket}
\title{RInputs_InternationalPrices_RegionMarket}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220 rows and 4 columns.
}
\usage{
RInputs_InternationalPrices_RegionMarket
}
\description{
RInputs_InternationalPrices_RegionMarket
}
\keyword{datasets}
