% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{RunCoreModel}
\alias{RunCoreModel}
\title{RunCoreModel}
\usage{
RunCoreModel(
  MainSegmentedDataTable_,
  MainPowerLargeDataTable_,
  BaseYear_,
  AnalysisEndYear_,
  CTRange_ = CTRange,
  DoPower_ = DoPower,
  NumberOfYears_ = NumberOfYears,
  RetirementProportion_ = RetirementProportion
)
}
\description{
Run Core Model
}
